Fix parameter parsing in `Vec3Unproject` / `Vec3UnprojectArray`
The float -> long -> float round trip looks suspicious, and may very well be the cause for https://ccpgames.atlassian.net/browse/PLAT-6839. Unfortunately there are no tests for this function, so it's a bit hard to tell whether this truly fixes the issue. But it is in line with other places where Python 3 is a lot more strict on type checking.